[FIX] non-standard usage of 't' in fopen mode
Problem: when opening or appending to files, sflfile.{c,h} uses a mode
of "wt" or "at". This is non-standard and in Cygwin causes '\n' to be
translated to "\r\n" in the output file. Unfortunately this makes the
generated code unsuitable for use as shell script.

Solution: stick to standard open modes.