[FIX] sflbits code unit bugs found by unit tests
Problem: the sflbits code unit contains bugs as shown by the unit tests.
These bugs are related to code with side effects embedded inside ASSERT
macros. These macros are replaced by no-op when compiling with the
macros NDEBUG defined as is the case for release and code-coverage.

Solution: move the side effect outside of the ASSERT and assert the
return code from the side-effect statement.