[sfl] Update ASSERT macro to remove unused-variable warnings
Problem: when in release or code-coverage mode the NDEBUG macro is set.
This causes all 'assert' macro to becom empty statement. If the argument
to an ASSERT was a variable the variable may then become set but unused.
This in turn causes a warning from the compiler.

Solution: use a different macro body depending on the fact NDEBUG is
defined or not. When it is, replace the 'assert' with a statement which
uses the argument but compiles to nothing.

References:
- https://stackoverflow.com/questions/777261/avoiding-unused-variables-warnings-when-using-assert-in-a-release-build
- http://cnicholson.net/2009/02/stupid-c-tricks-adventures-in-assert/