[FIX][tap] Fix usage of precision/field-width in printf
Problem: the tap library format comments so that each line starts with a
'#' comment marker. In order to do so it must print only part of the
comment up to an end-of-line. This substring printing is done by using a
precision flags in a printf call. Unfortunately instead of a precision
flag the code uses a minimum field-width flag. This means the whole
string is printed, not just the selected part.

Solution: use a correct precision flag (a period).