[FIX] Fix ASSERT related bugs in different code unit
Problem: many code units suffer from ASSERT related bugs (they use a
statement with side-effects inside the assert). When compiling with
NDEBUG defined the ASSERT disappear with the asserted code but the
asserted code is needed for correctness.

Solution: Extract the code from the ASSERT and use an integer variable
to hold the return value of this code. Assert the return value.