[project] Seach for libm in configure.ac
Problem: on certain platforms (Cygwin) there is no need to link with
libm but on others (Linux) it is needed. We should always search for the
library that provides the mathematical functions.

Solution: Use AC_SEARCH_LIBS to find the library providing 'cos()'. The
use of AC_SEARCH_LIBS, means that the library if found is added to the
global list of libraries for all link targets.