[project] Generate the main header files
Problem: we want to restore the project to the original state were most
files were genrated (either with libero or gsl).

Solution: start with the simplest files, the main library headers wich
are just concatenation of the individual headers in each library
directory.

Note: we have a problem with the 'dist' target. The .inc and .d files
are not included in the distribution tarball.