[project] Fix tests target in makefile.am
Problem: When creating the new automake infrastructure we removed the
check_PROGRAMS from the TESTS variables. Unfortunately only the programs
listed in the TESTS variables are used as test-suite. Also, distcheck
runs the whole testsuite so there is no need to run make check
separately.

Solution: update the product.am files to list the tests in the TESTS
variable. Update travis script to remove the 'make check' command.