[build] Add build variants to the build system
Problem: as a maintainer we often want to validate specific configure
options. The current build system doesn't allow to easily switch between
different sets of configure options.

Solution: Use a mechanism of 'variant' where each variant is a script
file defining the different tools and options we want to use at
configure and build time. Update the build script to use this mechanism.
Update the travis.yml to use the variant mechanism.